refactor(input-edit): clean up after initial review
- use memo when finding dataFormCells
- make arrow key navigation not depend on table size
- rename component folders to fit component names
