+++ packages/input-edit/src/main.js
@@ -5,7 +5,7 @@ import {hot} from 'react-hot-loader/root'
 import PropTypes from 'prop-types'
 
 import reducers, {sagas} from './modules/reducers'
-import InputEdit from './components/inputedit/InputEditContainer'
+import InputEdit from './components/InputEdit/InputEditContainer'
 
 const packageName = 'input-edit'
